refactor: isolate components from main
